#267 promote intbitset-based Span implementation as the one
